Changed the check password code to prioritise checking if password is common first. If it is then it will always be weak.
added in feature to check the pwned api when user adds their password to the vault and informs if their added password is breached.

updated the useful hints and tips tab to be better laid out and added a 4th tab that uses another free API from API Ninja that generates a secure password of any length. Trailing a clipboard feature that allows a user to copy the password and then after 15 seconds clears the clip board to prevent the plain text password remaining longer than it needs to be.